test: reduce test in debug: no package built, no cross-features crates tested

+++ .github/scripts/cargo_build.sh
@@ -10,24 +10,25 @@ set -ex
 
 ROOT_FOLDER=$(pwd)
 
-# First build the Debian and RPM packages. It must come at first since
-# after this step `ckms` and `cosmian_kms_server` are built with custom features flags (fips for example).
-rm -rf target/"$TARGET"/debian
-rm -rf target/"$TARGET"/generate-rpm
-
-if [ -f /etc/redhat-release ]; then
-  cd crate/cli && cargo build --target "$TARGET" --release && cd -
-  cd crate/server && cargo build --target "$TARGET" --release && cd -
-  cargo install --version 0.14.1 cargo-generate-rpm --force
-  cd "$ROOT_FOLDER"
-  cargo generate-rpm --target "$TARGET" -p crate/cli
-  cargo generate-rpm --target "$TARGET" -p crate/server --metadata-overwrite=pkg/rpm/scriptlets.toml
-elif [ -f /etc/lsb-release ]; then
-  cargo install --version 2.4.0 cargo-deb --force
-  cargo deb --target "$TARGET" -p cosmian_kms_cli --variant fips
-  cargo deb --target "$TARGET" -p cosmian_kms_cli
-  cargo deb --target "$TARGET" -p cosmian_kms_server --variant fips
-  cargo deb --target "$TARGET" -p cosmian_kms_server
+if [ "$DEBUG_OR_RELEASE" = "release" ]; then
+  # First build the Debian and RPM packages. It must come at first since
+  # after this step `ckms` and `cosmian_kms_server` are built with custom features flags (fips for example).
+  rm -rf target/"$TARGET"/debian
+  rm -rf target/"$TARGET"/generate-rpm
+  if [ -f /etc/redhat-release ]; then
+    cd crate/cli && cargo build --target "$TARGET" --release && cd -
+    cd crate/server && cargo build --target "$TARGET" --release && cd -
+    cargo install --version 0.14.1 cargo-generate-rpm --force
+    cd "$ROOT_FOLDER"
+    cargo generate-rpm --target "$TARGET" -p crate/cli
+    cargo generate-rpm --target "$TARGET" -p crate/server --metadata-overwrite=pkg/rpm/scriptlets.toml
+  elif [ -f /etc/lsb-release ]; then
+    cargo install --version 2.4.0 cargo-deb --force
+    cargo deb --target "$TARGET" -p cosmian_kms_cli --variant fips
+    cargo deb --target "$TARGET" -p cosmian_kms_cli
+    cargo deb --target "$TARGET" -p cosmian_kms_server --variant fips
+    cargo deb --target "$TARGET" -p cosmian_kms_server
+  fi
 fi
 
 if [ -z "$TARGET" ]; then
@@ -60,15 +61,18 @@ fi
 
 rustup target add "$TARGET"
 
-# Before building the crates, test crates individually on specific features
-cargo install --version 0.6.31 cargo-hack --force
-crates=("crate/kmip" "crate/client" "crate/cli" "crate/server")
-for crate in "${crates[@]}"; do
-  echo "cargo hack on $crate"
-  cd "$crate"
-  cargo hack check --feature-powerset --no-dev-deps
-  cd "$ROOT_FOLDER"
-done
+# Additional tests
+if [ "$DEBUG_OR_RELEASE" = "release" ]; then
+  # Before building the crates, test crates individually on specific features
+  cargo install --version 0.6.31 cargo-hack --force
+  crates=("crate/kmip" "crate/client" "crate/cli" "crate/server")
+  for crate in "${crates[@]}"; do
+    echo "cargo hack on $crate"
+    cd "$crate"
+    cargo hack check --feature-powerset --no-dev-deps
+    cd "$ROOT_FOLDER"
+  done
+fi
 
 echo "Building crate/pkcs11/provider"
 cd crate/pkcs11/provider
